+++ src/main/kotlin/com/github/lunakoly/quicklink/repository/RepositoryInfo.kt
@@ -1,6 +1,9 @@
 package com.github.lunakoly.quicklink.repository
 
+import com.intellij.openapi.vfs.VirtualFile
+
 class RepositoryInfo(
+    val root: VirtualFile,
     val branch: String?,
     val commitHash: String,
     val remotes: Map<String, String>,